+++ src/lib/Table/Header.js
@@ -1,24 +1,55 @@
-export default {
+const Cell = {
   functional: true,
   props: ['field'],
   render (h, { parent, props }) {
-    const orderedBy = parent.sortOrder[props.field.name]
-    const icon = <i staticClass="uk-icon-justify uk-margin-small-left" class={{
-      'uk-invisible': !orderedBy,
-      'vk-icon-arrow-down': orderedBy === 'asc' || orderedBy === undefined,
-      'vk-icon-arrow-up': orderedBy === 'desc'
-    }}></i>
-    return (
-      <th class={{
-        'uk-visible-hover-inline': props.field.sortBy,
-        'vk-table-order': props.field.sortBy,
+    const { field } = props
+    const { sortBy, headerClass } = field
+    const orderedBy = parent.sortOrder[field.name]
+    // Header, default or custom render
+    const Header = (typeof field.header === 'function')
+      ? h({
+        functional: true,
+        props: ['field'],
+        render: field.header
+      }, { props: { field } })
+      : field.header
+    return h('th', {
+      class: {
+        'uk-visible-hover-inline': sortBy,
+        'vk-table-order': sortBy,
         'uk-active': orderedBy,
-        [props.field.headerClass]: props.field.headerClass
-      }} on-click={e => props.field.sortBy && parent.emitSort(props.field)}>
-        <span class="uk-position-relative">
-          { props.field.header }{ props.field.sortBy && icon }
-        </span>
-      </th>
-    )
+        [headerClass]: headerClass
+      },
+      on: {
+        click: e => sortBy && parent.emitSort(props.field)
+      }
+    }, [
+      sortBy
+        ? h('span', {
+          class: 'uk-position-relative'
+        }, [
+          Header,
+          h(Icon, { props: { orderedBy } })
+        ])
+        : Header
+    ])
   }
 }
+
+const Icon = {
+  functional: true,
+  props: ['orderedBy'],
+  render (h, { props }) {
+    const { orderedBy } = props
+    return h('i', {
+      staticClass: 'uk-icon-justify uk-margin-small-left',
+      class: {
+        'uk-invisible': !orderedBy,
+        'vk-icon-arrow-down': orderedBy === 'asc' || orderedBy === undefined,
+        'vk-icon-arrow-up': orderedBy === 'desc'
+      }
+    })
+  }
+}
+
+export default Cell